backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic
pydantic-settings
python-multipart
httpx
python-dotenv
apscheduler
feedparser
beautifulsoup4
lxml
google-generativeai
pywebpush
cryptography
